User based authentication methods (#130)
* noot

* noot

* wip

* wip

* wip

* noot

* noot

* noot

* noot

* noot

* add some ai documentation

* multiple transports

* argon2 fix

* npm i

* change default token expiry

* noot

* fix build

* remove unused env var

* add some logging

* add GITLAB_API_URL to the examples

* argon2 warning

* Fix list of tools

* feat: Add NPM publish workflow for automated package publishing (#208)

* single logger

---------
