+++ src/customSchemas.ts
@@ -1,19 +1,5 @@
 import { z } from "zod";
-import { pino } from 'pino';
 const DEFAULT_NULL = process.env.DEFAULT_NULL === "true";
-
-
-const logger = pino({
-  level: process.env.LOG_LEVEL || 'info',
-  transport: {
-    target: 'pino-pretty',
-    options: {
-      colorize: true,
-      levelFirst: true,
-      destination: 2,
-    },
-  },
-});
 
 export const flexibleBoolean = z.preprocess(val => {
   if (typeof val === "boolean") {
