+++ src/logger.ts
@@ -1,0 +1,12 @@
+import { pino } from 'pino';
+
+export const logger = pino({
+  level: process.env.LOG_LEVEL || 'info',
+  transport: {
+    target: 'pino-pretty',
+    options: {
+      colorize: true,
+      levelFirst: true,
+    },
+  },
+});
